+++ aiccs-20191015/alibabacloud_aiccs20191015/models.py
@@ -20617,6 +20617,263 @@ class HotlineSessionQueryResponse(TeaModel):
         return self
 
 
+class ImportOneTaskPhoneNumberRequest(TeaModel):
+    def __init__(
+        self,
+        out_id: str = None,
+        owner_id: int = None,
+        phone_number: str = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
+        task_id: int = None,
+        variables: Dict[str, Any] = None,
+    ):
+        self.out_id = out_id
+        self.owner_id = owner_id
+        # This parameter is required.
+        self.phone_number = phone_number
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
+        # This parameter is required.
+        self.task_id = task_id
+        self.variables = variables
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.out_id is not None:
+            result['OutId'] = self.out_id
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
+        if self.phone_number is not None:
+            result['PhoneNumber'] = self.phone_number
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
+        if self.task_id is not None:
+            result['TaskId'] = self.task_id
+        if self.variables is not None:
+            result['Variables'] = self.variables
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('OutId') is not None:
+            self.out_id = m.get('OutId')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
+        if m.get('PhoneNumber') is not None:
+            self.phone_number = m.get('PhoneNumber')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
+        if m.get('TaskId') is not None:
+            self.task_id = m.get('TaskId')
+        if m.get('Variables') is not None:
+            self.variables = m.get('Variables')
+        return self
+
+
+class ImportOneTaskPhoneNumberShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        out_id: str = None,
+        owner_id: int = None,
+        phone_number: str = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
+        task_id: int = None,
+        variables_shrink: str = None,
+    ):
+        self.out_id = out_id
+        self.owner_id = owner_id
+        # This parameter is required.
+        self.phone_number = phone_number
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
+        # This parameter is required.
+        self.task_id = task_id
+        self.variables_shrink = variables_shrink
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.out_id is not None:
+            result['OutId'] = self.out_id
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
+        if self.phone_number is not None:
+            result['PhoneNumber'] = self.phone_number
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
+        if self.task_id is not None:
+            result['TaskId'] = self.task_id
+        if self.variables_shrink is not None:
+            result['Variables'] = self.variables_shrink
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('OutId') is not None:
+            self.out_id = m.get('OutId')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
+        if m.get('PhoneNumber') is not None:
+            self.phone_number = m.get('PhoneNumber')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
+        if m.get('TaskId') is not None:
+            self.task_id = m.get('TaskId')
+        if m.get('Variables') is not None:
+            self.variables_shrink = m.get('Variables')
+        return self
+
+
+class ImportOneTaskPhoneNumberResponseBodyData(TeaModel):
+    def __init__(
+        self,
+        detail_id: int = None,
+    ):
+        self.detail_id = detail_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.detail_id is not None:
+            result['DetailId'] = self.detail_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('DetailId') is not None:
+            self.detail_id = m.get('DetailId')
+        return self
+
+
+class ImportOneTaskPhoneNumberResponseBody(TeaModel):
+    def __init__(
+        self,
+        access_denied_detail: str = None,
+        code: str = None,
+        data: ImportOneTaskPhoneNumberResponseBodyData = None,
+        message: str = None,
+        request_id: str = None,
+        success: bool = None,
+    ):
+        self.access_denied_detail = access_denied_detail
+        self.code = code
+        self.data = data
+        self.message = message
+        self.request_id = request_id
+        self.success = success
+
+    def validate(self):
+        if self.data:
+            self.data.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.access_denied_detail is not None:
+            result['AccessDeniedDetail'] = self.access_denied_detail
+        if self.code is not None:
+            result['Code'] = self.code
+        if self.data is not None:
+            result['Data'] = self.data.to_map()
+        if self.message is not None:
+            result['Message'] = self.message
+        if self.request_id is not None:
+            result['RequestId'] = self.request_id
+        if self.success is not None:
+            result['Success'] = self.success
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('AccessDeniedDetail') is not None:
+            self.access_denied_detail = m.get('AccessDeniedDetail')
+        if m.get('Code') is not None:
+            self.code = m.get('Code')
+        if m.get('Data') is not None:
+            temp_model = ImportOneTaskPhoneNumberResponseBodyData()
+            self.data = temp_model.from_map(m['Data'])
+        if m.get('Message') is not None:
+            self.message = m.get('Message')
+        if m.get('RequestId') is not None:
+            self.request_id = m.get('RequestId')
+        if m.get('Success') is not None:
+            self.success = m.get('Success')
+        return self
+
+
+class ImportOneTaskPhoneNumberResponse(TeaModel):
+    def __init__(
+        self,
+        headers: Dict[str, str] = None,
+        status_code: int = None,
+        body: ImportOneTaskPhoneNumberResponseBody = None,
+    ):
+        self.headers = headers
+        self.status_code = status_code
+        self.body = body
+
+    def validate(self):
+        if self.body:
+            self.body.validate()
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.headers is not None:
+            result['headers'] = self.headers
+        if self.status_code is not None:
+            result['statusCode'] = self.status_code
+        if self.body is not None:
+            result['body'] = self.body.to_map()
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('headers') is not None:
+            self.headers = m.get('headers')
+        if m.get('statusCode') is not None:
+            self.status_code = m.get('statusCode')
+        if m.get('body') is not None:
+            temp_model = ImportOneTaskPhoneNumberResponseBody()
+            self.body = temp_model.from_map(m['body'])
+        return self
+
+
 class ImportTaskNumberDatasRequest(TeaModel):
     def __init__(
         self,
@@ -25994,6 +26251,7 @@ class QueryAiCallDetailPageRequest(TeaModel):
         batch_id: str = None,
         call_result: str = None,
         called_number: str = None,
+        detail_ids: List[int] = None,
         end_calling_time: int = None,
         end_imported_time: int = None,
         major_intent: str = None,
@@ -26013,6 +26271,7 @@ class QueryAiCallDetailPageRequest(TeaModel):
         self.batch_id = batch_id
         self.call_result = call_result
         self.called_number = called_number
+        self.detail_ids = detail_ids
         self.end_calling_time = end_calling_time
         self.end_imported_time = end_imported_time
         self.major_intent = major_intent
@@ -26046,6 +26305,8 @@ class QueryAiCallDetailPageRequest(TeaModel):
             result['CallResult'] = self.call_result
         if self.called_number is not None:
             result['CalledNumber'] = self.called_number
+        if self.detail_ids is not None:
+            result['DetailIds'] = self.detail_ids
         if self.end_calling_time is not None:
             result['EndCallingTime'] = self.end_calling_time
         if self.end_imported_time is not None:
@@ -26086,6 +26347,145 @@ class QueryAiCallDetailPageRequest(TeaModel):
             self.call_result = m.get('CallResult')
         if m.get('CalledNumber') is not None:
             self.called_number = m.get('CalledNumber')
+        if m.get('DetailIds') is not None:
+            self.detail_ids = m.get('DetailIds')
+        if m.get('EndCallingTime') is not None:
+            self.end_calling_time = m.get('EndCallingTime')
+        if m.get('EndImportedTime') is not None:
+            self.end_imported_time = m.get('EndImportedTime')
+        if m.get('MajorIntent') is not None:
+            self.major_intent = m.get('MajorIntent')
+        if m.get('MaxConversationDuration') is not None:
+            self.max_conversation_duration = m.get('MaxConversationDuration')
+        if m.get('MinConversationDuration') is not None:
+            self.min_conversation_duration = m.get('MinConversationDuration')
+        if m.get('OutId') is not None:
+            self.out_id = m.get('OutId')
+        if m.get('OwnerId') is not None:
+            self.owner_id = m.get('OwnerId')
+        if m.get('PageNo') is not None:
+            self.page_no = m.get('PageNo')
+        if m.get('PageSize') is not None:
+            self.page_size = m.get('PageSize')
+        if m.get('ResourceOwnerAccount') is not None:
+            self.resource_owner_account = m.get('ResourceOwnerAccount')
+        if m.get('ResourceOwnerId') is not None:
+            self.resource_owner_id = m.get('ResourceOwnerId')
+        if m.get('StartCallingTime') is not None:
+            self.start_calling_time = m.get('StartCallingTime')
+        if m.get('StartImportedTime') is not None:
+            self.start_imported_time = m.get('StartImportedTime')
+        if m.get('Status') is not None:
+            self.status = m.get('Status')
+        if m.get('TaskId') is not None:
+            self.task_id = m.get('TaskId')
+        return self
+
+
+class QueryAiCallDetailPageShrinkRequest(TeaModel):
+    def __init__(
+        self,
+        batch_id: str = None,
+        call_result: str = None,
+        called_number: str = None,
+        detail_ids_shrink: str = None,
+        end_calling_time: int = None,
+        end_imported_time: int = None,
+        major_intent: str = None,
+        max_conversation_duration: int = None,
+        min_conversation_duration: int = None,
+        out_id: str = None,
+        owner_id: int = None,
+        page_no: int = None,
+        page_size: int = None,
+        resource_owner_account: str = None,
+        resource_owner_id: int = None,
+        start_calling_time: int = None,
+        start_imported_time: int = None,
+        status: int = None,
+        task_id: str = None,
+    ):
+        self.batch_id = batch_id
+        self.call_result = call_result
+        self.called_number = called_number
+        self.detail_ids_shrink = detail_ids_shrink
+        self.end_calling_time = end_calling_time
+        self.end_imported_time = end_imported_time
+        self.major_intent = major_intent
+        self.max_conversation_duration = max_conversation_duration
+        self.min_conversation_duration = min_conversation_duration
+        self.out_id = out_id
+        self.owner_id = owner_id
+        self.page_no = page_no
+        self.page_size = page_size
+        self.resource_owner_account = resource_owner_account
+        self.resource_owner_id = resource_owner_id
+        self.start_calling_time = start_calling_time
+        self.start_imported_time = start_imported_time
+        # This parameter is required.
+        self.status = status
+        # This parameter is required.
+        self.task_id = task_id
+
+    def validate(self):
+        pass
+
+    def to_map(self):
+        _map = super().to_map()
+        if _map is not None:
+            return _map
+
+        result = dict()
+        if self.batch_id is not None:
+            result['BatchId'] = self.batch_id
+        if self.call_result is not None:
+            result['CallResult'] = self.call_result
+        if self.called_number is not None:
+            result['CalledNumber'] = self.called_number
+        if self.detail_ids_shrink is not None:
+            result['DetailIds'] = self.detail_ids_shrink
+        if self.end_calling_time is not None:
+            result['EndCallingTime'] = self.end_calling_time
+        if self.end_imported_time is not None:
+            result['EndImportedTime'] = self.end_imported_time
+        if self.major_intent is not None:
+            result['MajorIntent'] = self.major_intent
+        if self.max_conversation_duration is not None:
+            result['MaxConversationDuration'] = self.max_conversation_duration
+        if self.min_conversation_duration is not None:
+            result['MinConversationDuration'] = self.min_conversation_duration
+        if self.out_id is not None:
+            result['OutId'] = self.out_id
+        if self.owner_id is not None:
+            result['OwnerId'] = self.owner_id
+        if self.page_no is not None:
+            result['PageNo'] = self.page_no
+        if self.page_size is not None:
+            result['PageSize'] = self.page_size
+        if self.resource_owner_account is not None:
+            result['ResourceOwnerAccount'] = self.resource_owner_account
+        if self.resource_owner_id is not None:
+            result['ResourceOwnerId'] = self.resource_owner_id
+        if self.start_calling_time is not None:
+            result['StartCallingTime'] = self.start_calling_time
+        if self.start_imported_time is not None:
+            result['StartImportedTime'] = self.start_imported_time
+        if self.status is not None:
+            result['Status'] = self.status
+        if self.task_id is not None:
+            result['TaskId'] = self.task_id
+        return result
+
+    def from_map(self, m: dict = None):
+        m = m or dict()
+        if m.get('BatchId') is not None:
+            self.batch_id = m.get('BatchId')
+        if m.get('CallResult') is not None:
+            self.call_result = m.get('CallResult')
+        if m.get('CalledNumber') is not None:
+            self.called_number = m.get('CalledNumber')
+        if m.get('DetailIds') is not None:
+            self.detail_ids_shrink = m.get('DetailIds')
         if m.get('EndCallingTime') is not None:
             self.end_calling_time = m.get('EndCallingTime')
         if m.get('EndImportedTime') is not None:
@@ -34680,6 +35080,7 @@ class UpdateAiCallTaskRequest(TeaModel):
         resource_owner_account: str = None,
         resource_owner_id: int = None,
         start_type: str = None,
+        task_cps: int = None,
         task_id: str = None,
         task_name: str = None,
         task_start_time: int = None,
@@ -34698,6 +35099,8 @@ class UpdateAiCallTaskRequest(TeaModel):
         self.resource_owner_id = resource_owner_id
         # This parameter is required.
         self.start_type = start_type
+        # This parameter is required.
+        self.task_cps = task_cps
         # This parameter is required.
         self.task_id = task_id
         # This parameter is required.
@@ -34735,6 +35138,8 @@ class UpdateAiCallTaskRequest(TeaModel):
             result['ResourceOwnerId'] = self.resource_owner_id
         if self.start_type is not None:
             result['StartType'] = self.start_type
+        if self.task_cps is not None:
+            result['TaskCps'] = self.task_cps
         if self.task_id is not None:
             result['TaskId'] = self.task_id
         if self.task_name is not None:
@@ -34767,6 +35172,8 @@ class UpdateAiCallTaskRequest(TeaModel):
             self.resource_owner_id = m.get('ResourceOwnerId')
         if m.get('StartType') is not None:
             self.start_type = m.get('StartType')
+        if m.get('TaskCps') is not None:
+            self.task_cps = m.get('TaskCps')
         if m.get('TaskId') is not None:
             self.task_id = m.get('TaskId')
         if m.get('TaskName') is not None:
@@ -34791,6 +35198,7 @@ class UpdateAiCallTaskShrinkRequest(TeaModel):
         resource_owner_account: str = None,
         resource_owner_id: int = None,
         start_type: str = None,
+        task_cps: int = None,
         task_id: str = None,
         task_name: str = None,
         task_start_time: int = None,
@@ -34809,6 +35217,8 @@ class UpdateAiCallTaskShrinkRequest(TeaModel):
         self.resource_owner_id = resource_owner_id
         # This parameter is required.
         self.start_type = start_type
+        # This parameter is required.
+        self.task_cps = task_cps
         # This parameter is required.
         self.task_id = task_id
         # This parameter is required.
@@ -34846,6 +35256,8 @@ class UpdateAiCallTaskShrinkRequest(TeaModel):
             result['ResourceOwnerId'] = self.resource_owner_id
         if self.start_type is not None:
             result['StartType'] = self.start_type
+        if self.task_cps is not None:
+            result['TaskCps'] = self.task_cps
         if self.task_id is not None:
             result['TaskId'] = self.task_id
         if self.task_name is not None:
@@ -34878,6 +35290,8 @@ class UpdateAiCallTaskShrinkRequest(TeaModel):
             self.resource_owner_id = m.get('ResourceOwnerId')
         if m.get('StartType') is not None:
             self.start_type = m.get('StartType')
+        if m.get('TaskCps') is not None:
+            self.task_cps = m.get('TaskCps')
         if m.get('TaskId') is not None:
             self.task_id = m.get('TaskId')
         if m.get('TaskName') is not None: